Handle non-arrays for the parseOrderByPhrase method
The adapter allows sending an object for the `parseOrderByPhrase`
method, but it does not handle the case properly. This is the case
when trying to skip elements but not providing a sort order.

The fix consists of converting the sort object into an array, if it
is not already one.